refactor: clean up AlpacaProvider implementation and remove unused provider module
